Change to double

+++ src/app.scala
@@ -8,23 +8,25 @@ object app {
     println(s"Value e^5 ==> ${eX(5.0000, 9, 1)}")
     println(s"Value e^5 ==> ${eX(0.5000, 9, 1)}")
     println(s"Value e^5 ==> ${eX(-0.5000, 9, 1)}")
+
+
   }
 
-  def eX(x:Float, count:Int, acc:Float):Float = count match{
+  def eX(x:Double, count:Int, acc:Double):Double = count match{
     case c if c == 1 => x + acc
     case c if c > 1 => eX(x, count-1, acc + div (x, count))
   }
 
-  def div(x:Float, i:Int):Float = {
+  def div(x:Double, i:Int):Double = {
     pow(x,i,1) / factorial(i,1)
   }
 
-  def factorial(n:Int, acc:Int):Float = n match {
+  def factorial(n:Int, acc:Int):Double = n match {
     case x if x == 1 => acc
     case x if x > 1 => factorial( n - 1, n * acc)
   }
 
-  def pow(x:Float, pot:Int, acc:Float):Float = pot match {
+  def pow(x:Double, pot:Int, acc:Double):Double = pot match {
     case n if n == 0 => acc
     case n if n > 0 => pow(x, pot-1, acc * x)
   }
